refactor: moved more components after further study of codebase
