+++ src/components/Molecules/JSONTranslationArea.js
@@ -1,5 +1,5 @@
 import {Container} from "@material-ui/core";
-import JSONCreator from "../Molecules/JSONCreator";
+import JSONCreator from "./JSONCreator";
 import React from "react";
 
 export const JSONTranslationArea = () => {
